support for new image upload widget and updates
these updates have been tested to not break any functionality, and QR codes still work. ncu wants to "update" vue-qrcode-reader to 3.1.0, which is actually a downgrade from the current prelease version in use which is newer and working. The ncu suggestion was held back for this package because of this.