fix migations and backend, change climb to a radio selection
This should get the backend most if not all of the way there. I came back today wanting to do something, but I forgot exactly what. Pushing entries works fine on my end so this should work for now.

I changed the migration so that it drops the existing table before creating a new one. If this isn't desired, you can rename the old table manually or we can change it to modify the table instead of deleting and recreating it.

I fixed the API Schema by comparing the request the frontend was sending to what the validator was expecting and saw that the app combined the is_blue, ds_position, and team_number fields into one team_data string. I also added some code to process this new string so that these properties can be accessed on their own without string processing.

I also changed the climb to use a radio button as a robot will only do one climb (or no climb) in a match. This makes storing the value a lot easier as you can just use an integer.